Add files via upload
Firstly, I addressed the issue of repeated cards by introducing a set data structure to keep track of used cards during the dealing process in the Player class. This ensures that each card dealt is unique, preventing duplicates within the player's hand. Also, now, face cards are correctly labeled as "A" for Ace, "J" for Jack, "Q" for Queen, and "K" for King, which was removed originally to test functionality.