Add files via upload
I fixed an issue where the player's total was unintentionally affected during the dealer's card draw phase. The problem originated from the fact that the addCard function, used for both the player and the dealer, not only added a card to the dealer's hand but also updated the player's total. To resolve this, I introduced a new function called addCardToDealer specifically designed for the dealer. This function takes care of adding a card to the dealer's hand without altering the player's total.